added jwt authentication and authorization

+++ routers/posts.py
@@ -7,6 +7,7 @@ from sqlalchemy.orm import selectinload
 import models
 from db import get_db
 from schemas import PostCreate, PostResponse, PostUpdate
+from auth import get_current_user
 
 router = APIRouter(
     prefix="/api/posts",
@@ -33,13 +34,13 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
     return post
 
 @router.post("", response_model=PostResponse, status_code=status.HTTP_201_CREATED)
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def create_post(post: PostCreate, current_user:Annotated[models.User, Depends(get_current_user)],db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.User).where(models.User.id == post.user_id))
     user = result.scalars().first()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
 
-    new_post = models.Post(title=post.title, content=post.content, user_id=post.user_id)
+    new_post = models.Post(title=post.title, content=post.content, user_id=current_user.id)
     db.add(new_post)
     await db.commit()
     await db.refresh(new_post, attribute_names=["author"])
@@ -65,11 +66,20 @@ async def patch_post(
     return post
 
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id: int, current_user:Annotated[models.User, Depends(get_current_user)],db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+    
+
+
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="you can delete only your posts"
+        )
+
     await db.delete(post)
     await db.commit()
     return None
